changes to turbidityNorm function (all channels)

+++ src/sensors/SpectralModule.h
@@ -1,13 +1,13 @@
 #include "DFRobot_AS7341.h"
 
 struct spectralChannels {
-	uint16_t f1 = 0;
-	uint16_t f2 = 0;
-	uint16_t f3 = 0;
-	uint16_t f4 = 0;
-	uint16_t f5 = 0;
-	uint16_t f6 = 0;
-	uint16_t f7 = 0;
+	float f1 = 0;
+	float f2 = 0;
+	float f3 = 0;
+	float f4 = 0;
+	float f5 = 0;
+	float f6 = 0;
+	float f7 = 0;
 	float f8 = 0;
 	float nir = 0;
 };
